% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENphylo_modeling.R
\name{ENphylo_modeling}
\alias{ENphylo_modeling}
\title{Calculating species marginality and specialization via ENFA and
 phylogenetic imputation}
\usage{
ENphylo_modeling(input_data, tree, input_mask, obs_col, time_col=NULL,
 min_occ_enfa=50, boot_test_perc=20, boot_reps=10, swap.args= list(nsim=10,
 si=0.2, si2=0.2), eval.args=list(eval_metric_for_imputation="AUC",
 eval_threshold=0.7,output_options="best"),clust=0.5,output.dir='.')
}
\arguments{
\item{input_data}{a list of \code{sf::data.frame} objects containing species
occurrence data in binary format (ones for presence, zero for background
points) along with the explanatory variables to be used in ENFA or ENphylo.
Each element of the list must be named (i.e. the species name).
Alternatively, ENFA model outputs generated through \code{ENphylo_modeling}
can be supplied as named elements of \code{input_data} list.}

\item{tree}{an object of class \code{phylo} including all the species in
\code{input_data}. The tree needs not to be ultrametric or fully
dichotomous.}

\item{input_mask}{an object of class \code{SpatRaster}. This is the
geographical mask defining the spatial domain encompassing the background
area enclosing all the species in the tree.}

\item{obs_col}{character. Name of the \code{input_data} column containing the
vector of species occurrence data in binary format.}

\item{time_col}{character. Name of the \code{input_data} column containing the
time intervals associated to each species presence and background point
(optional).}

\item{min_occ_enfa}{numeric. The minimum number of occurrence data required
for a species to be modeled with ENFA.}

\item{boot_test_perc}{numeric. Percentage of data (ranging between 0 and 100)
used to calibrate ENFA and/or ENphylo models within a bootstrap
cross-validation scheme. The \code{100-boot_test_perc} remaining percentage
will be used for evaluating model performances.}

\item{boot_reps}{numeric. Number of evaluation runs performed within the
bootstrap cross-validation scheme to evaluate ENFA and/or ENphylo models. If
set to 0, models evaluation is skipped and the internal evaluation element
returns \code{NULL}.}

\item{swap.args}{list of ENphylo parameters. It includes: \itemize{ \item nsim
= number of alternative phylogenies (altered topology and branch lengths as
compared to the reference tree) to be generated by means of
\code{\link[RRphylo]{swapONE}} and tested (\code{nsim < 1} is not allowed -
see details); \item si,si2 = arguments passed to \code{RRphylo::swapONE}.}}

\item{eval.args}{list of evaluation model parameters. It includes:
\itemize{\item eval_metric_for_imputation = evaluation metric used to select
the most accurate ENphylo models. The viable options are: \code{"AUC"},
\code{"TSS"}, \code{"CBI"}, \code{"SORENSEN"}, or \code{"OMR"};
\item eval_threshold = the minimum evaluation score to evaluate ENFA and ENphylo
performance. ENFA models having \code{eval_metric_for_imputation} lower than
\code{eval_threshold} are compared to ENphylo models to keep the one fitting
best. Additionally, within ENphylo, models derived from the swapped trees
having \code{eval_metric_for_imputation} lower than \code{eval_threshold}
are excluded from the output; \item output_options = the strategy adopted to
return ENphylo models results (see details). The viable options are:
\code{"full"}, \code{"weighted.mean"}, and \code{"best"}.
}}

\item{clust}{numeric. The proportion of cores used to train ENFA and ENphylo
models. If \code{NULL}, parallel computing is disabled. It is set at 0.5 by
default.}

\item{output.dir}{the file path wherein \code{ENphylo_modeling} creates
"ENphylo_enfa_models" and "ENphylo_imputed_models"" folders to store
modeling outputs (see details).}
}
\value{
A list of length equal to the number of modeled species. For each
 element of this list, the output contains three elements: \enumerate{ \item
 \strong{$call} a character specifying the algorithm used to model the
 species (i.e. ENFA or ENphylo). \item \strong{$formatted data} a list of
 input data formatted to run either ENFA or ENphylo algorithms. Specifically,
 the list reports: the presence data points (\code{$input_ones}),
 the background points (\code{$input_back}),the name
 of the columns associated to the arguments \code{OBS_col} and
 \code{time_col} (if specified), the name of the column containing the cell
 numbers (\code{geoID_col}), and the coordinates
 of presence data only (\code{$one_coords}). \item \strong{$calibrated_model}
 a list of three elements. The output objects are different depending on
 whether ENFA or ENphylo is used to model the species:}

ENFA

\itemize{\item$call: a character specifying the algorithm used.
 \item$full_ model: a list containing marginality and specialization factors,
 the 'co' matrix, the number of significant axes, and all the other objects
 generated by applying ENFA on the entire occurrence dataset (see
 \cite{Rinnan et al. 2019} for additional details). \item$evaluation: a
 matrix containing the evaluation scores of the ENFA model assessed by all
 possible evaluation metrics (i.e. Area Under the Curve (AUC), True Skill
 Statistic (TSS), Boyce Index (CBI), Sorensen Index, and Omission Rate (OMR))
 for each model evaluations run.}

ENphylo

\itemize{\item$call: a character specifying the algorithm used.
 \item$co: a list of the 'co' matrices of length equal to the number of
 alternative phylogenies tested (i.e. \code{nsim} argument).
 \item$evaluation: a data.frame containing the evaluation scores of ENphylo
 model assessed by all possible evaluation metrics for each alternative
 phylogeny. The output of this object depends on the strategy adopted by the
 user through the \code{output_options} argument (see
 details).\item$output_options:  a character vector including the argument
 \code{output_options} and \code{eval_metric_for_imputation} set to run the
 of ENphylo model.}
}
\description{
The function computes vectors of marginality and specialization
 according to \cite{Rinnan & Lawler (2019)} via Environmental Niche Factor
 Analysis (ENFA) and phylogenetic imputation (\cite{Garland & Ives, 2000}).
 It takes a list of \code{Spatial} or \code{Simple Features} (\pkg{sp} or
 \pkg{sf}) objects and a phylogenetic tree to train ENFA and/or ENphylo
 models. Both model techniques are calibrated and evaluated while accounting
 for phylogenetic uncertainty. Calibrations are made on a random subset of
 the data under the bootstrap cross-validation scheme. The predictive power
 of the different models is estimated using four different evaluation
 metrics.
}
\details{
\code{ENphylo_modeling} automatically arranges \code{input_data} in a
 convenient format to run ENFA or ENphylo. The internal call of the function
 is \code{"calibrated_enfa"} for ENFA and \code{"calibrated_imputed"} for
 ENphylo, respectively.

\strong{Phylogenetic uncertainty}

 The function cannot work with \code{nsim} < 1 since one of the strongest
 points of \code{ENphylo_modeling} is to test alternative phylogenies to
 provide the most accurate possible reconstruction of species environmental
 preferences.

\strong{Phylogenetic Imputation}

 \code{ENphylo_modeling} automatically switches from ENFA to ENphylo
 algorithm for any species having less than \code{min_occ_enfa} occurrences
 or ENFA model accuracy below \code{eval_threshold}. In this latter case, the
 function performs both models and retains the one performing best according
 to \code{eval_metric_for_imputation}. Phylogenetic imputation is allowed for
 up to 30\% of the species on the tree. If the number of species to impute
 exceeds 30\%, \code{ENphylo_modeling} automatically splits the original tree
 into smaller subtrees, so that the maximum percentage of imputation is
 observed. Each subtree is designed as to impute phylogenetically distant
 species and to retain species phylogenetically close to the taxa to be
 imputed (so that imputation is robust). In this case, the functions prints
 the number of phylogenies used.

\strong{Outputs}

 If \code{ENphylo_modeling} runs the ENphylo algorithm, the outputs depends
 on the strategy adopted by the user through the \code{output_options}
 argument. If \code{output_options="full"}, all CO matrices and evaluation
 metrics for all the swapped trees are returned. Under
 \code{output_options="weighted.mean"}, the output consists of a subset of CO
 matrices and evaluation metrics for those tree swapping iterations achieving
 a predictive accuracy in terms of \code{eval_metric_for_imputation} above
 \code{eval_threshold}. Finally, if \code{output_options="best"}, a single CO
 matrix and evaluation scores vector corresponding to the most accurate
 swapped tree is returned. If any of the tree swapping iterations under
 either \code{"best"} or \code{"weighted.mean"} provides below-the-threshold
 accuracy, the function automatically switches to \code{"full"} strategy.

 Eventually, the function creates two new folders, "ENphylo_enfa_models" and
 "ENphylo_imputed_models", in \code{output.dir}. In each of these folders, a
 number of new named folders equal to the number of modeled species are
 created. Therein, model outputs and background area in .tif format are saved
 as \code{model_outputs.RData} and \code{study_area}, respectively.
}
\examples{
\dontrun{
setwd("YOUR_DIRECTORY")
getwd()->main.dir
url<-"https://www.dropbox.com/s/wm7qcuwm0nm35w2/ENphylo\%20code\%26data\%202.zip?dl=1"
download.file(url,file.path(main.dir,"ENphylo code&data 2.zip"),mode="wb")
unzip("ENphylo code&data 2.zip")
load("ENphylo code&data/example_data.RData")
 ### NOTE: the object enfa_all_species is no longer necessary with RRgeo,
 ### it can be removed from the example environment

MASK_FULL<-terra::rast("ENphylo code&data/variable_bio1.tif")
external_data<-terra::rast(list.files("ENphylo code&data/external_data",full.names=TRUE))

## NOTE: Given the size of the data, running the function is time-comsuming
### CASE 1
ENmod<-ENphylo_modeling(input_data=DATA_FULL,
                        tree=tree,
                        input_mask=MASK_FULL,
                        obs_col="OBS",
                        time_col="TIME_factor",
                        eval.args=list(eval_metric_for_imputation = "AUC",
                                       output_options = "best"))
gc()

### CASE 2
## Simulating a rare species by subsampling its occurrences to force run ENphylo algorithm
k=10 # index of the species to subsample as in DATA_FULL (k available 1-31)
npoints=10 # Number of randomly selected presence data points for the k species in DATA_FULL

# Subsampling procedure
Species_data_start<-DATA_FULL[[k]]
Species_data1_start<-subset(Species_data_start, OBS==1)
Species_data0_start<-subset(Species_data_start, OBS==0)
sample(nrow(Species_data1_start),npoints)->zz
Species_data1<-Species_data1_start[zz,]
Species_data0<-subset(Species_data0_start,TIME_factor\%in\%Species_data1$TIME_factor)
Species_data<-rbind(Species_data1,Species_data0)
Species_data->DATA_FULL[[k]]

# Running ENphylo
ENmod_subsam<-ENphylo_modeling(input_data=DATA_FULL,
                               tree=tree,
                               input_mask=MASK_FULL,
                               obs_col="OBS",
                               time_col="TIME_factor",
                               min_occ_enfa=npoints+1,
                               eval.args=list(eval_metric_for_imputation="AUC",
                                              eval_threshold=0.7,
                                              output_options = "best"))
gc()
}
}
\references{
Rinnan, D. S., &  Lawler, J. (2019). Climate-niche factor
 analysis: a spatial approach to quantifying species vulnerability to climate
 change. \emph{Ecography}, 42(9), 1494–1503. doi/full/10.1111/ecog.03937

Garland, T., & Ives, A. R. (2000). Using the past to predict the
 present: Confidence intervals for regression equations in phylogenetic
 comparative methods. \emph{American Naturalist}, 155(3),346–364.
 doi.org/10.1086/303327
}
\author{
Alessandro Mondanaro, Mirko Di Febbraro, Silvia Castiglione, Carmela
 Serio, Marina Melchionna, Pasquale Raia
}
