% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENphylo_prediction.R
\name{ENphylo_prediction}
\alias{ENphylo_prediction}
\title{Project the ENFA and ENphylo models into new geographical space and
 time interval}
\usage{
ENphylo_prediction(object, newdata, calc_method=c("raster", "terra"),
 convert.to.suitability=FALSE)
}
\arguments{
\item{object}{a \code{list} returned by the \code{\link{ENphylo_modeling}}
function where all the species are modelled with ENFA or ENphylo models.}

\item{newdata}{a set of explanatory variables onto which ENFA or ENphylo
models will be projected. It could be a \code{data.frame} or a
\code{rasterStack} object.}

\item{calc_method}{character. The R package and related functions which
\code{ENphylo_prediction} uses to obtain model predictions.}

\item{convert.to.suitability}{logical. If \code{TRUE},
\code{ENphylo_prediction} projects ENFA or ENphylo model predictions in
other areas and other time scales.}
}
\value{
A list of length equal to the number of modelled species. For each
 element of the list, the output objects differ depending on whether ENFA or
 ENphylo is implemented and if a raster or data.frame object is provided by
 the user as new environments where to project the model. Furthermore, if the
 \code{convert.to.suitability} is set to \code{TRUE}, each element of the
 list includes the habitat suitability values in addition to model
 predictions for marginality and specialization. In any case, each element of
 the output list contains two elements:

ENFA/raster

\enumerate{ \item\strong{$call}: a character specifying the algorithm
 used for the model predictions. \item\strong{$enfa_prediction}: a list of
 RasterLayer objects including marginality and specialization model
 predictions plus the habitat suitability and binary presence/absence maps
 obtained by using three different predefined thresholds: MaxSensSpec (i.e.
 maximize TSS), SensSpec (i.e. equalize sensitivity and specificity) and 10th
 percentile of predicted probability. Habitat suitability estimates and
 binary values are calculated only if \code{convert.to.suitability = TRUE}.}

ENFA/data.frame

\enumerate{ \item\strong{$call}: a character specifying the algorithm
 used for the model predictions.\item\strong{$enfa_prediction}: a data.frame
 containing marginality and specialization model predictions and the habitat
 suitability values (only if \code{convert.to.suitability = TRUE}).}

ENphylo/raster

\enumerate{ \item\strong{$call}: a character specifying the algorithm
 used for the model predictions. \item\strong{$imputed_prediction}: a list of
 raster objects including marginality and specialization model predictions.
 The layer number of each raster is equal to the integer number chosen for
 testing alternative phylogenies (i.e. \code{nsim} argument). If
 \code{convert.to.suitability = TRUE}, in addition to marginality and
 specialization, the function provides raster objects containing habitat
 suitability and binary presence/absence maps (see above) for each tested
 phylogeny.}

ENphylo/data.frame

\enumerate{ \item\strong{$call}: a character specifying the algorithm
 used for the model predictions. \item\strong{$imputed_prediction}: a list of
 data.frame objects of length equal to the integer number chosen for testing
 alternative phylogenies (i.e. \code{nsim} argument) containing marginality
 and specialization model predictions and the habitat suitability values
 (only if \code{convert.to.suitability argument = TRUE}).}
}
\description{
The function projects species marginality and specialization
 factors in other areas and other time scales. The function is able to
 convert marginality and specialization factors in habitat suitability values
 by using the Mahalanobis distances method.
}
\details{
If \code{convert.to.suitability} is set as \code{TRUE},
 \code{ENphylo_prediction} uses the function
 \code{\link[adehabitatHS]{mahasuhab}} from the \pkg{adehabitatHS} R
 package (\cite{Calenge, 2006}) to compute the habitat suitability map of the
 species over a given area. The conversion of Mahalanobis distances into
 probabilities follows the chi-squared distribution. Specifically, we set the
 degree of freedom equal to \emph{n} rather than \emph{n-1} following
 \cite{Etherington (2019)}. To convert habitat suitability values into binary
 presence/absence values, \code{ENphylo_prediction} relies on
 \code{\link{optimal.thresholds}} function in the package
 \pkg{PresenceAbsence} (\cite{Freeman & Moisen, 2008}).
}
\examples{
\dontrun{
setwd("YOUR_DIRECTORY")
getwd()->main.dir
url<-"https://www.dropbox.com/s/wm7qcuwm0nm35w2/ENphylo\%20code\%26data\%202.zip?dl=1"
download.file(url,file.path(main.dir,"ENphylo code&data 2.zip"),mode="wb")
unzip("ENphylo code&data 2.zip")
load("ENphylo code&data/example_data.RData")
MASK_FULL<-raster::raster("ENphylo code&data/variable_bio1.tif")
external_data<-raster::stack(list.files("ENphylo code&data/external_data",full.names=TRUE))

## NOTE: Given the size of the data, running the function is time-comsuming
### CASE 1
ENmod<-ENphylo_modeling(input_data=DATA_FULL,
                       tree=tree,
                       input_mask=MASK_FULL,
                       obs_col="OBS",
                       time_col="TIME_factor",
                       eval_metric_for_imputation = "AUC",
                       output_options = "best")
gc()

# Predicting for the first species in ENmod output (available 1-31)
ENmod_pred<-ENphylo_prediction(object=ENmod[1],
                              newdata=external_data,
                              convert.to.suitability=TRUE,
                              calc_method="terra")

### CASE 2
## Simulating a rare species by subsampling its occurrences to force run ENphylo algorithm
k=10 # index of the species to subsample as in DATA_FULL (k available 1-31)
npoints=10 # Number of randomly selected presence data points for the k species in DATA_FULL

# Subsampling procedure
Species_data_start<-DATA_FULL[[k]]
Species_data1_start<-subset(Species_data_start, OBS==1)
Species_data0_start<-subset(Species_data_start, OBS==0)
sample(nrow(Species_data1_start),npoints)->zz
Species_data1<-Species_data1_start[zz,]
Species_data0<-subset(Species_data0_start,TIME_factor\%in\%Species_data1$TIME_factor)
Species_data<-rbind(Species_data1,Species_data0)
Species_data->DATA_FULL[[k]]

# Running ENphylo
ENmod_subsam<-ENphylo_modeling(input_data=DATA_FULL,
                              tree=tree,
                              input_mask=MASK_FULL,
                              obs_col="OBS",
                              time_col="TIME_factor",
                              min_occ_enfa=npoints+1,
                              eval_metric_for_imputation="AUC",
                              eval_threshold=0.7,
                              output_options = "best")
gc()

# Predicting for the k species in ENmod_subsam output
ENmod_subsam_pred<-ENphylo_prediction(object=ENmod_subsam[k],
                                     newdata=external_data,
                                     convert.to.suitability=TRUE,
                                     calc_method="terra")


### CASE 3
## In case you do not trust an ENFA-derived SDM (regardless of the evaluation score)
## you might want to run ENphylo anyway, retrieving the ENFA models and specifing
## species k you what to look at ENphylo prediction for

# Extract a species modelled with ENFA under ENphylo_modeling and 'force' ENphylo calibration on it
sample(which(sapply(ENmod,function(x) x$call)=="calibrated_enfa"),1)->k


ENmod3<-ENphylo_modeling(external_enfa_models=ENmod,
                             spec_for_imputation=names(ENmod)[k],
                             tree=tree,
                             output_options="full",
                             eval_metric_for_imputation="AUC")

gc()

# Transform the raster object to a data.frame object to predict on data.frame
external_dataframe<-raster::rasterToPoints(external_data)[,-c(1,2)]

# Predicting for the k species in ENmod_enfa1 output
ENmod3_pred<-ENphylo_prediction(object=ENmod3[k],
                                    newdata=external_dataframe,
                                    convert.to.suitability=TRUE,
                                    calc_method="raster")


### CASE 4
# Run ENphylo under alternative model evaluation metric and strategy
ENmod4<-ENphylo_modeling(external_enfa_models=ENmod,
                             spec_for_imputation=names(ENmod)[k],
                             tree=tree,
                             output_options="best",
                             eval_metric_for_imputation="TSS",
                             eval_threshold = 0.4)
gc()

# Predicting for the k species in ENmod_enfa2 output
ENmod4_pred<-ENphylo_prediction(object=ENmod4[k],
                                    newdata=external_dataframe,
                                    convert.to.suitability=TRUE,
                                    calc_method="terra")
}
}
\references{
Calenge, C. (2006) The package adehabitat for the R software: a
 tool for the analysis of space and habitat use by animals. \emph{Ecological
 Modelling}, 197, 516-519.

Etherington, T. R. (2019). Mahalanobis distances and ecological
 niche modelling: correcting a chi-squared probability error. \emph{PeerJ},
 7, e6678.

Freeman, E. A. & Moisen, G. (2008). PresenceAbsence: An R Package
 for Presence-Absence Model Analysis. \emph{Journal of Statistical Software},
 23(11):1-31.
}
\author{
Alessandro Mondanaro, Mirko Di Febbraro, Silvia Castiglione,
 Carmela Serio, Marina Melchionna, Pasquale Raia
}
